+++ rs/dogecoin/ckdoge/minter/src/lib.rs
@@ -3,11 +3,13 @@ mod tests;
 
 pub mod address;
 pub mod candid_api;
+pub mod fees;
 pub mod lifecycle;
 pub mod updates;
 
 use crate::address::DogecoinAddress;
 use crate::dogecoin_canister::MillikoinuPerByte;
+use crate::fees::DogecoinFeeEstimator;
 use crate::lifecycle::init::Network;
 use async_trait::async_trait;
 use candid::Principal;
@@ -40,6 +42,12 @@ pub struct DogeCanisterRuntime {}
 
 #[async_trait]
 impl CanisterRuntime for DogeCanisterRuntime {
+    type Estimator = DogecoinFeeEstimator;
+
+    fn fee_estimator(&self, state: &CkBtcMinterState) -> DogecoinFeeEstimator {
+        DogecoinFeeEstimator::from_state(state)
+    }
+
     fn refresh_fee_percentiles_frequency(&self) -> Duration {
         const SIX_MINUTES: Duration = Duration::from_secs(360);
         SIX_MINUTES