+++ rs/pocket_ic_server/src/pocket_ic.rs
@@ -12,7 +12,8 @@ use axum::{
     extract::State,
     response::{Html, IntoResponse},
 };
-use bitcoin::Network as BtcAdapterNetwork;
+use bitcoin::Network as BitcoinAdapterNetwork;
+use bitcoin::dogecoin::Network as DogecoinAdapterNetwork;
 use bytes::Bytes;
 use candid::{CandidType, Decode, Encode, Principal};
 use cycles_minting_canister::{
@@ -26,8 +27,10 @@ use hyper::header::{CONTENT_TYPE, HeaderValue};
 use hyper::{Method, StatusCode};
 use ic_boundary::{Health, RootKey, status};
 use ic_btc_adapter::config::{Config as BitcoinAdapterConfig, IncomingSource as BtcIncomingSource};
-use ic_btc_adapter::start_server as start_btc_server;
-use ic_btc_interface::{Fees as BtcFees, InitConfig as BtcInitConfig, Network as BtcNetwork};
+use ic_btc_adapter::{AdapterNetwork, start_server as start_btc_server};
+use ic_btc_interface::{
+    Fees as BitcoinFees, InitConfig as BitcoinInitConfig, Network as BitcoinNetwork,
+};
 use ic_config::adapters::AdaptersConfig;
 use ic_config::execution_environment::MAX_CANISTER_HTTP_REQUESTS_IN_FLIGHT;
 use ic_config::{
@@ -35,6 +38,9 @@ use ic_config::{
     subnet_config::SubnetConfig,
 };
 use ic_crypto_sha2::Sha256;
+use ic_doge_interface::{
+    Fees as DogecoinFees, InitConfig as DogecoinInitConfig, Network as DogecoinNetwork,
+};
 use ic_http_endpoints_public::query;
 use ic_http_endpoints_public::{
     CanisterReadStateServiceBuilder, IngressValidatorBuilder, QueryServiceBuilder,
@@ -58,19 +64,20 @@ use ic_interfaces_state_manager::StateReader;
 use ic_limits::MAX_P2P_IO_CHANNEL_SIZE;
 use ic_logger::{ReplicaLogger, no_op_logger};
 use ic_management_canister_types_private::{
-    BoundedVec, CanisterIdRecord, CanisterInstallMode, CanisterSettingsArgs, EcdsaCurve,
-    EcdsaKeyId, LogVisibilityV2, MasterPublicKeyId, Method as Ic00Method,
-    ProvisionalCreateCanisterWithCyclesArgs, SchnorrAlgorithm, SchnorrKeyId, VetKdCurve,
-    VetKdKeyId,
+    BoundedVec, CanisterIdRecord, CanisterInstallMode, CanisterSettingsArgs,
+    CanisterSnapshotDataKind, EcdsaCurve, EcdsaKeyId, LogVisibilityV2, MasterPublicKeyId,
+    Method as Ic00Method, ProvisionalCreateCanisterWithCyclesArgs, ReadCanisterSnapshotDataArgs,
+    ReadCanisterSnapshotMetadataArgs, SchnorrAlgorithm, SchnorrKeyId, VetKdCurve, VetKdKeyId,
 };
 use ic_metrics::MetricsRegistry;
 use ic_nervous_system_common::ONE_YEAR_SECONDS;
 use ic_nns_common::types::UpdateIcpXdrConversionRatePayload;
 use ic_nns_constants::{
     BITCOIN_TESTNET_CANISTER_ID, CYCLES_LEDGER_CANISTER_ID, CYCLES_LEDGER_INDEX_CANISTER_ID,
-    CYCLES_MINTING_CANISTER_ID, GOVERNANCE_CANISTER_ID, IDENTITY_CANISTER_ID, LEDGER_CANISTER_ID,
-    LEDGER_INDEX_CANISTER_ID, LIFELINE_CANISTER_ID, MIGRATION_CANISTER_ID, NNS_UI_CANISTER_ID,
-    REGISTRY_CANISTER_ID, ROOT_CANISTER_ID, SNS_AGGREGATOR_CANISTER_ID, SNS_WASM_CANISTER_ID,
+    CYCLES_MINTING_CANISTER_ID, DOGECOIN_CANISTER_ID, GOVERNANCE_CANISTER_ID, IDENTITY_CANISTER_ID,
+    LEDGER_CANISTER_ID, LEDGER_INDEX_CANISTER_ID, LIFELINE_CANISTER_ID, MIGRATION_CANISTER_ID,
+    NNS_UI_CANISTER_ID, REGISTRY_CANISTER_ID, ROOT_CANISTER_ID, SNS_AGGREGATOR_CANISTER_ID,
+    SNS_WASM_CANISTER_ID,
 };
 use ic_nns_delegation_manager::{NNSDelegationBuilder, NNSDelegationReader};
 use ic_nns_governance_api::{NetworkEconomics, Neuron, neuron::DissolveState};
@@ -102,7 +109,8 @@ use ic_types::batch::BlockmakerMetrics;
 use ic_types::ingress::{IngressState, IngressStatus};
 use ic_types::messages::{CertificateDelegationFormat, CertificateDelegationMetadata};
 use ic_types::{
-    CanisterId, Cycles, Height, NodeId, NumInstructions, PrincipalId, RegistryVersion, SubnetId,
+    CanisterId, Cycles, Height, NodeId, NumInstructions, PrincipalId, RegistryVersion, SnapshotId,
+    SubnetId,
     artifact::UnvalidatedArtifactMutation,
     canister_http::{
         CanisterHttpReject, CanisterHttpRequest as AdapterCanisterHttpRequest,
@@ -138,10 +146,10 @@ use std::hash::Hash;
 use std::str::FromStr;
 use std::{
     collections::{BTreeMap, BTreeSet, HashMap},
-    fs::{File, remove_file},
+    fs::{File, OpenOptions, remove_file},
     io::{BufReader, Read, Write},
     net::SocketAddr,
-    path::PathBuf,
+    path::{Path, PathBuf},
     sync::{Arc, Mutex, RwLock},
     time::{Duration, SystemTime, UNIX_EPOCH},
 };
@@ -186,6 +194,7 @@ const INTERNET_IDENTITY_TEST_CANISTER_WASM: &[u8] =
 const NNS_DAPP_TEST_CANISTER_WASM: &[u8] = include_bytes!(env!("NNS_DAPP_TEST_CANISTER_WASM_PATH"));
 const BITCOIN_TESTNET_CANISTER_WASM: &[u8] =
     include_bytes!(env!("BITCOIN_TESTNET_CANISTER_WASM_PATH"));
+const DOGECOIN_CANISTER_WASM: &[u8] = include_bytes!(env!("DOGECOIN_CANISTER_WASM_PATH"));
 const MIGRATION_CANISTER_WASM: &[u8] = include_bytes!(env!("MIGRATION_CANISTER_WASM_PATH"));
 
 const DEFAULT_SUBACCOUNT: Subaccount = Subaccount([0; 32]);
@@ -200,7 +209,7 @@ const INITIAL_CYCLES: u128 = u128::MAX / 2;
 // - it is still possible to top up the canister with further cycles without overflowing 64-bit range.
 const INITIAL_CYCLES_64_BIT: u64 = u64::MAX / 2;
 
-// Maximum duration of waiting for bitcoin/canister http adapter server to start.
+// Maximum duration of waiting for bitcoin/dogecoin/canister http adapter server to start.
 const MAX_START_SERVER_DURATION: Duration = Duration::from_secs(60);
 
 // Clippy complains that these are interior-mutable.
@@ -208,6 +217,9 @@ const MAX_START_SERVER_DURATION: Duration = Duration::from_secs(60);
 // https://rust-lang.github.io/rust-clippy/master/index.html#/declare_interior_mutable_const
 #[allow(clippy::declare_interior_mutable_const)]
 const CONTENT_TYPE_CBOR: HeaderValue = HeaderValue::from_static("application/cbor");
+
+// Maximum data chunk size when downloading/uploading canister snapshots.
+const MAX_CHUNK_SIZE: u64 = 2_000_000;
 
 fn default_timestamp(icp_features: &Option<IcpFeatures>) -> SystemTime {
     // To set the ICP/XDR conversion rate, the PocketIC time (in seconds) must be strictly larger than the default timestamp in CMC state.
@@ -321,6 +333,7 @@ impl BitcoinAdapterParts {
     fn new(
         bitcoind_addr: Vec<SocketAddr>,
         uds_path: PathBuf,
+        network: AdapterNetwork,
         log_level: Option<Level>,
         replica_logger: ReplicaLogger,
         metrics_registry: MetricsRegistry,
@@ -332,7 +345,7 @@ impl BitcoinAdapterParts {
             logger: logger_config_from_level(log_level),
             incoming_source: BtcIncomingSource::Path(uds_path.clone()),
             address_limits: (1, 1),
-            ..BitcoinAdapterConfig::default_with(BtcAdapterNetwork::Regtest.into())
+            ..BitcoinAdapterConfig::default_with(network)
         };
         let adapter = tokio::spawn(async move {
             start_btc_server(replica_logger, metrics_registry, bitcoin_adapter_config).await
@@ -525,12 +538,14 @@ struct PocketIcSubnets {
     icp_config: IcpConfig,
     log_level: Option<Level>,
     bitcoind_addr: Option<Vec<SocketAddr>>,
+    dogecoind_addr: Option<Vec<SocketAddr>>,
     icp_features: Option<IcpFeatures>,
     initial_time: SystemTime,
     auto_progress_enabled: bool,
     gateway_port: Option<u16>,
     synced_registry_version: RegistryVersion,
     _bitcoin_adapter_parts: Option<BitcoinAdapterParts>,
+    _dogecoin_adapter_parts: Option<BitcoinAdapterParts>,
 }
 
 impl PocketIcSubnets {
@@ -545,6 +560,7 @@ impl PocketIcSubnets {
         icp_config: &IcpConfig,
         log_level: Option<Level>,
         bitcoin_adapter_uds_path: Option<PathBuf>,
+        dogecoin_adapter_uds_path: Option<PathBuf>,
     ) -> StateMachineBuilder {
         let subnet_type = conv_type(subnet_kind);
         let subnet_size = subnet_size(subnet_kind);
@@ -635,6 +651,7 @@ impl PocketIcSubnets {
             .with_registry_data_provider(registry_data_provider.clone())
             .with_log_level(log_level)
             .with_bitcoin_testnet_uds_path(bitcoin_adapter_uds_path)
+            .with_dogecoin_testnet_uds_path(dogecoin_adapter_uds_path)
             .create_at_registry_version(create_at_registry_version)
     }
 
@@ -644,6 +661,7 @@ impl PocketIcSubnets {
         icp_config: IcpConfig,
         log_level: Option<Level>,
         bitcoind_addr: Option<Vec<SocketAddr>>,
+        dogecoind_addr: Option<Vec<SocketAddr>>,
         icp_features: Option<IcpFeatures>,
         initial_time: SystemTime,
         auto_progress_enabled: bool,
@@ -672,12 +690,14 @@ impl PocketIcSubnets {
             icp_config,
             log_level,
             bitcoind_addr,
+            dogecoind_addr,
             icp_features,
             initial_time,
             auto_progress_enabled,
             gateway_port,
             synced_registry_version,
             _bitcoin_adapter_parts: None,
+            _dogecoin_adapter_parts: None,
         }
     }
 
@@ -763,6 +783,12 @@ impl PocketIcSubnets {
             } else {
                 None
             };
+        let dogecoin_adapter_uds_path =
+            if matches!(subnet_kind, SubnetKind::Bitcoin) && self.dogecoind_addr.is_some() {
+                Some(NamedTempFile::new().unwrap().into_temp_path().to_path_buf())
+            } else {
+                None
+            };
 
         let latest_registry_version = self.registry_data_provider.latest_version().get();
         let create_at_registry_version = if update_registry_and_system_canisters {
@@ -785,6 +811,7 @@ impl PocketIcSubnets {
             &self.icp_config,
             self.log_level,
             bitcoin_adapter_uds_path.clone(),
+            dogecoin_adapter_uds_path.clone(),
         );
 
         if let Some(subnet_id) = subnet_id {
@@ -863,11 +890,22 @@ impl PocketIcSubnets {
         }
 
         // We need `StateMachine` components (metrics/logger)
-        // to create a bitcoin adapter (if applicable).
+        // to create a bitcoin/dogecoin adapter (if applicable).
         if let Some(bitcoin_adapter_uds_path) = bitcoin_adapter_uds_path {
             self._bitcoin_adapter_parts = Some(BitcoinAdapterParts::new(
                 self.bitcoind_addr.clone().unwrap(),
                 bitcoin_adapter_uds_path,
+                AdapterNetwork::Bitcoin(BitcoinAdapterNetwork::Regtest),
+                self.log_level,
+                sm.replica_logger.clone(),
+                sm.metrics_registry.clone(),
+            ));
+        }
+        if let Some(dogecoin_adapter_uds_path) = dogecoin_adapter_uds_path {
+            self._dogecoin_adapter_parts = Some(BitcoinAdapterParts::new(
+                self.dogecoind_addr.clone().unwrap(),
+                dogecoin_adapter_uds_path,
+                AdapterNetwork::Dogecoin(DogecoinAdapterNetwork::Regtest),
                 self.log_level,
                 sm.replica_logger.clone(),
                 sm.metrics_registry.clone(),
@@ -945,6 +983,7 @@ impl PocketIcSubnets {
                 ii,
                 nns_ui,
                 bitcoin,
+                dogecoin,
                 canister_migration,
             } = icp_features;
             if let Some(ref config) = registry {
@@ -973,6 +1012,9 @@ impl PocketIcSubnets {
             }
             if let Some(ref config) = bitcoin {
                 self.deploy_bitcoin(config);
+            }
+            if let Some(ref config) = dogecoin {
+                self.deploy_dogecoin(config);
             }
             if let Some(ref config) = canister_migration {
                 self.deploy_canister_migration(config);
@@ -2154,9 +2196,9 @@ impl PocketIcSubnets {
             assert_eq!(canister_id, BITCOIN_TESTNET_CANISTER_ID);
 
             // Install the Bitcoin testnet canister.
-            let args = BtcInitConfig {
-                network: Some(BtcNetwork::Regtest),
-                fees: Some(BtcFees::testnet()),
+            let args = BitcoinInitConfig {
+                network: Some(BitcoinNetwork::Regtest),
+                fees: Some(BitcoinFees::testnet()),
                 ..Default::default()
             };
             btc_subnet
@@ -2165,6 +2207,79 @@ impl PocketIcSubnets {
                     canister_id,
                     CanisterInstallMode::Install,
                     BITCOIN_TESTNET_CANISTER_WASM.to_vec(),
+                    Encode!(&args).unwrap(),
+                )
+                .unwrap();
+        }
+    }
+
+    fn deploy_dogecoin(&self, config: &IcpFeaturesConfig) {
+        // Using a match here to force an update after changing
+        // the type of `IcpFeaturesConfig`.
+        match config {
+            IcpFeaturesConfig::DefaultConfig => (),
+        };
+
+        // Nothing to do if the Bitcoin subnet does not exist (yet).
+        let Some(ref btc_subnet) = self.btc_subnet else {
+            return;
+        };
+
+        if !btc_subnet
+            .state_machine
+            .canister_exists(DOGECOIN_CANISTER_ID)
+        {
+            // Create the Dogecoin mainnet canister with its ICP mainnet settings and configured for the regtest network.
+            // These settings have been obtained by calling
+            // `dfx canister call r7inp-6aaaa-aaaaa-aaabq-cai canister_status '(record {canister_id=principal"gordg-fyaaa-aaaan-aaadq-cai";})' --ic`:
+            //     settings = record {
+            //       freezing_threshold = opt (2_592_000 : nat);
+            //       wasm_memory_threshold = opt (0 : nat);
+            //       controllers = vec {
+            //         principal "r7inp-6aaaa-aaaaa-aaabq-cai";
+            //         principal "gordg-fyaaa-aaaan-aaadq-cai";
+            //       };
+            //       reserved_cycles_limit = opt (5_000_000_000_000 : nat);
+            //       log_visibility = opt variant { controllers };
+            //       wasm_memory_limit = opt (3_221_225_472 : nat);
+            //       memory_allocation = opt (0 : nat);
+            //       compute_allocation = opt (0 : nat);
+            //     };
+
+            let settings = CanisterSettingsArgs {
+                controllers: Some(BoundedVec::new(vec![
+                    ROOT_CANISTER_ID.get(),
+                    DOGECOIN_CANISTER_ID.get(),
+                ])),
+                compute_allocation: Some(0_u64.into()),
+                memory_allocation: Some(0_u64.into()),
+                freezing_threshold: Some(2_592_000_u64.into()),
+                reserved_cycles_limit: Some(5_000_000_000_000_u128.into()),
+                log_visibility: Some(LogVisibilityV2::Controllers),
+                log_memory_limit: Some(4_096_u64.into()),
+                wasm_memory_limit: Some(3_221_225_472_u64.into()),
+                wasm_memory_threshold: Some(0_u64.into()),
+                environment_variables: None,
+            };
+            let canister_id = btc_subnet.state_machine.create_canister_with_cycles(
+                Some(DOGECOIN_CANISTER_ID.get()),
+                Cycles::zero(),
+                Some(settings),
+            );
+            assert_eq!(canister_id, DOGECOIN_CANISTER_ID);
+
+            // Install the Dogecoin mainnet canister configured for the regtest network.
+            let args = DogecoinInitConfig {
+                network: Some(DogecoinNetwork::Regtest),
+                fees: Some(DogecoinFees::testnet()),
+                ..Default::default()
+            };
+            btc_subnet
+                .state_machine
+                .install_wasm_in_mode(
+                    canister_id,
+                    CanisterInstallMode::Install,
+                    DOGECOIN_CANISTER_WASM.to_vec(),
                     Encode!(&args).unwrap(),
                 )
                 .unwrap();
@@ -2451,6 +2566,7 @@ impl PocketIc {
         }
     }
 
+    #[allow(clippy::too_many_arguments)]
     pub(crate) fn try_new(
         runtime: Arc<Runtime>,
         seed: u64,
@@ -2459,6 +2575,7 @@ impl PocketIc {
         icp_config: IcpConfig,
         log_level: Option<Level>,
         bitcoind_addr: Option<Vec<SocketAddr>>,
+        dogecoind_addr: Option<Vec<SocketAddr>>,
         icp_features: Option<IcpFeatures>,
         incomplete_state: Option<IncompleteStateFlag>,
         initial_time: Option<Time>,
@@ -2716,6 +2833,7 @@ impl PocketIc {
             icp_config,
             log_level,
             bitcoind_addr,
+            dogecoind_addr,
             icp_features,
             initial_time,
             auto_progress_enabled,
@@ -3470,6 +3588,218 @@ impl Operation for AdvanceTimeAndTick {
 }
 
 #[derive(Clone, Debug)]
+pub struct CanisterSnapshotDownload {
+    pub sender: PrincipalId,
+    pub canister_id: CanisterId,
+    pub snapshot_id: SnapshotId,
+    pub snapshot_dir: PathBuf,
+}
+
+fn ensure_empty_dir(path: &Path) -> Result<(), String> {
+    // Create the directory if needed (including parents).
+    // Succeeds silently if the directory already exists.
+    std::fs::create_dir_all(path)
+        .map_err(|e| format!("Could not create snapshot directory: {}", e))?;
+
+    // Now ensure it's actually a directory.
+    if !path.is_dir() {
+        return Err("Snapshot directory path exists but is not a directory".to_string());
+    }
+
+    // Check if it is empty.
+    if std::fs::read_dir(path)
+        .map_err(|e| format!("Could not read snapshot directory: {}", e))?
+        .next()
+        .is_some()
+    {
+        return Err("Snapshot directory is not empty".to_string());
+    }
+
+    Ok(())
+}
+
+enum BlobKind {
+    WasmModule,
+    WasmMemory,
+    StableMemory,
+}
+
+impl BlobKind {
+    fn description(&self) -> &str {
+        match self {
+            BlobKind::WasmModule => "WASM module",
+            BlobKind::WasmMemory => "WASM memory",
+            BlobKind::StableMemory => "stable memory",
+        }
+    }
+}
+
+fn download_blob_to_file(
+    subnet: Arc<StateMachine>,
+    canister_id: CanisterId,
+    snapshot_id: SnapshotId,
+    blob_kind: BlobKind,
+    length: u64,
+    file: PathBuf,
+) -> Result<(), String> {
+    let mut offset = 0;
+    let mut file = OpenOptions::new()
+        .create(true) // create the file if it doesn't exist
+        .append(true) // allow appending
+        .open(file)
+        .map_err(|e| format!("Could not create {} file: {}", blob_kind.description(), e))?;
+    while offset < length {
+        let chunk_size = std::cmp::min(length - offset, MAX_CHUNK_SIZE);
+        let kind = match blob_kind {
+            BlobKind::WasmModule => CanisterSnapshotDataKind::WasmModule {
+                offset,
+                size: chunk_size,
+            },
+            BlobKind::WasmMemory => CanisterSnapshotDataKind::WasmMemory {
+                offset,
+                size: chunk_size,
+            },
+            BlobKind::StableMemory => CanisterSnapshotDataKind::StableMemory {
+                offset,
+                size: chunk_size,
+            },
+        };
+        let data_args = ReadCanisterSnapshotDataArgs {
+            canister_id: canister_id.into(),
+            snapshot_id,
+            kind,
+        };
+        let data = subnet
+            .read_canister_snapshot_data(&data_args)
+            .map_err(|e| e.description().to_string())?
+            .chunk;
+        file.write_all(&data)
+            .map_err(|e| format!("Could not write {} file: {}", blob_kind.description(), e))?;
+        offset += chunk_size;
+    }
+    Ok(())
+}
+
+impl Operation for CanisterSnapshotDownload {
+    fn compute(&self, pic: &mut PocketIc) -> OpOut {
+        let effective_principal = EffectivePrincipal::CanisterId(self.canister_id);
+        let subnet = route(pic, effective_principal, false);
+        match subnet {
+            Ok(subnet) => {
+                if let Err(e) = ensure_empty_dir(&self.snapshot_dir) {
+                    return OpOut::Error(PocketIcError::InvalidCanisterSnapshotDirectory(e));
+                }
+
+                // Download snapshot metadata.
+                let metadata_args = ReadCanisterSnapshotMetadataArgs {
+                    canister_id: self.canister_id.into(),
+                    snapshot_id: self.snapshot_id,
+                };
+                let metadata = match subnet.read_canister_snapshot_metadata(&metadata_args) {
+                    Ok(metadata) => metadata,
+                    Err(e) => {
+                        return OpOut::Error(PocketIcError::CanisterSnapshotError(
+                            e.description().to_string(),
+                        ));
+                    }
+                };
+                let metadata_bytes = serde_json::to_string_pretty(&metadata).unwrap();
+                let metadata_path = self.snapshot_dir.join("metadata.json");
+                if let Err(e) = std::fs::write(metadata_path, metadata_bytes) {
+                    return OpOut::Error(PocketIcError::CanisterSnapshotError(format!(
+                        "Could not write metadata file: {}",
+                        e
+                    )));
+                }
+
+                // Download WASM binary.
+                let wasm_module_path = self.snapshot_dir.join("wasm_module.bin");
+                if let Err(e) = download_blob_to_file(
+                    subnet.clone(),
+                    self.canister_id,
+                    self.snapshot_id,
+                    BlobKind::WasmModule,
+                    metadata.wasm_module_size,
+                    wasm_module_path,
+                ) {
+                    return OpOut::Error(PocketIcError::CanisterSnapshotError(e));
+                }
+
+                // Download WASM memory.
+                let wasm_memory_path = self.snapshot_dir.join("wasm_memory.bin");
+                if let Err(e) = download_blob_to_file(
+                    subnet.clone(),
+                    self.canister_id,
+                    self.snapshot_id,
+                    BlobKind::WasmMemory,
+                    metadata.wasm_memory_size,
+                    wasm_memory_path,
+                ) {
+                    return OpOut::Error(PocketIcError::CanisterSnapshotError(e));
+                }
+
+                // Download stable memory.
+                if metadata.stable_memory_size != 0 {
+                    let stable_memory_path = self.snapshot_dir.join("stable_memory.bin");
+                    if let Err(e) = download_blob_to_file(
+                        subnet.clone(),
+                        self.canister_id,
+                        self.snapshot_id,
+                        BlobKind::StableMemory,
+                        metadata.stable_memory_size,
+                        stable_memory_path,
+                    ) {
+                        return OpOut::Error(PocketIcError::CanisterSnapshotError(e));
+                    }
+                }
+
+                // Download WASM chunk store.
+                let chunk_store = match subnet.get_snapshot_chunk_store(&metadata_args) {
+                    Ok(chunk_store) => chunk_store,
+                    Err(e) => {
+                        return OpOut::Error(PocketIcError::CanisterSnapshotError(
+                            e.description().to_string(),
+                        ));
+                    }
+                };
+                if !chunk_store.is_empty() {
+                    let chunk_store_path = self.snapshot_dir.join("wasm_chunk_store");
+                    if let Err(e) = std::fs::create_dir_all(&chunk_store_path) {
+                        return OpOut::Error(PocketIcError::CanisterSnapshotError(format!(
+                            "Could not create WASM chunk store directory: {}",
+                            e
+                        )));
+                    }
+                    for (hash, chunk) in chunk_store {
+                        let hash_str = hex::encode(hash);
+                        let chunk_file = chunk_store_path.join(format!("{hash_str}.bin"));
+                        if let Err(e) = std::fs::write(chunk_file, chunk) {
+                            return OpOut::Error(PocketIcError::CanisterSnapshotError(format!(
+                                "Could not write WASM chunk: {}",
+                                e
+                            )));
+                        }
+                    }
+                }
+
+                OpOut::NoOutput
+            }
+            Err(e) => OpOut::Error(PocketIcError::CanisterRequestRoutingError(e)),
+        }
+    }
+
+    fn id(&self) -> OpId {
+        OpId(format!(
+            "canister_snapshot_download(sender={},canister_id={},snapshot_id={},snapshot_dir='{}')",
+            self.sender,
+            self.canister_id,
+            self.snapshot_id,
+            self.snapshot_dir.display()
+        ))
+    }
+}
+
+#[derive(Clone, Debug)]
 pub struct SubmitIngressMessage(pub CanisterCall);
 
 impl Operation for SubmitIngressMessage {
@@ -3818,6 +4148,9 @@ impl Operation for CallRequest {
         match subnet {
             Err(e) => OpOut::Error(PocketIcError::CanisterRequestRoutingError(e)),
             Ok(subnet) => {
+                // Make sure the latest state is certified for the ingress filter to work.
+                subnet.certify_latest_state();
+
                 let node = &subnet.nodes[0];
                 let (s, mut r) = mpsc::channel(MAX_P2P_IO_CHANNEL_SIZE);
                 let ingress_filter = subnet.ingress_filter.clone();
@@ -4721,6 +5054,7 @@ mod tests {
                 None,
                 None,
                 None,
+                None,
                 false,
                 None,
             )
@@ -4734,6 +5068,7 @@ mod tests {
                 },
                 None,
                 IcpConfig::default(),
+                None,
                 None,
                 None,
                 None,